improves the folder organization structure a bit
