--- public/debug-panel.js
@@ -1,6 +1,0 @@
-import { StacheElement, type, ObservableObject, stache } from "./can.js";
-
-
-class DebugPanel extends StacheElement {
-    static view = ``;
-}
