+++ public/jira-oidc-helpers/fetch.ts
@@ -1,7 +1,7 @@
 /**
  * this module contains helper functions for requesting json over http. 
  */
-import {responseToJSON} from "../shared/response-to-json";
+import {responseToJSON} from "../utils/fetch/response-to-json";
 import { JsonResponse } from "../shared/types";
 
 export function nativeFetchJSON<T>(url: string, options?: RequestInit): Promise<JsonResponse<T>> {
