+++ public/reports/status-report.js
@@ -2,7 +2,7 @@ import { StacheElement, type, ObservableObject, stache } from "../can.js";
 
 import { dateFormatter } from "../issue-tooltip.js";
 
-import { DAY_IN_MS } from "../date-helpers.js";
+import { DAY_IN_MS } from "../utils/date/date-helpers.js";
 
 import { showTooltip } from "../issue-tooltip.js";
 import { workTypes } from "../jira/derived/work-status/work-status";
