+++ public/request-helpers/hosted-request-helper.js
@@ -1,4 +1,4 @@
-import {responseToJSON} from "../shared/response-to-json";
+import {responseToJSON} from "../utils/fetch/response-to-json";
 
 function fetchFromLocalStorage(key) {
   return window.localStorage.getItem(key);
