+++ public/shared/README.MD
@@ -1,0 +1,6 @@
+This folder should eventually be removed.
+
+
+It was intended to house modules shared between the autoscheduler and the timeline report.
+
+As that sort of code sharing is unlikely anymore, these modules should probably be organized by some other pattern.
