+++ public/timeline-configuration/timeline-configuration.js
@@ -1,11 +1,7 @@
 import { StacheElement, type, ObservableObject, ObservableArray, value, queues } from "../can.js";
 
 import { saveJSONToUrl, updateUrlParam } from "../shared/state-storage.js";
-// import {
-//   calculationKeysToNames,
-//   allTimingCalculationOptions,
-//   getImpliedTimingCalculations,
-// } from "../prepare-issues/date-data.js";
+
 
 import { createRoot } from "react-dom/client";
 import { createElement } from "react";
